Refresh_LINUX1.0.0.py
As it was introduced as RMB32 into DDE in 2014 last, as it was my collage finale year project. Currently I do enhance it into this handler pack for make a better utility. Hope it works.
ALl packages & program are platform independent but user need to use PYTHON 2.x compiler for smooth utility.